collector/requirements.txt
flask
requests
google-cloud-bigquery
